test for isSubdivided before calculating the odd vertices
test for isSubdivided before calculating the odd vertices. Currently,
only for boundary vertices. Still need to implement for interior
vertices.
Note: Is it worth it? The code is getting too complicated. Will be a
bitch to troubleshoot if something goes wrong.
Also, we are storing the oddvertices on an neighbor basis. This cause
duplicates. Perhaps there is a way to minimize this by keeping a global
array of odd vertices and only keeping the index to that array in the
odd_vertices list in the vertex.